+++ lib/src/decimal_format_util.dart
@@ -21,14 +21,15 @@ String formatNumber(
   int precision = 0, // 展示精度
   bool showSign = false, //是否展示符号位+
   RoundMode? mode,
-  bool cutInvalidZero = false, //删除尾部零.
-  bool showCompact = false, // 是否压缩大数展示(T, B, M, K)
+  bool cutInvalidZero = false, //删除尾部零
+  bool showCompact = false, // 是否使用精简转换器转换大数展示, 优先于千分位展示
+  CompactConverter? compactConverter, // 自定义精简转换器
   bool showThousands = false, // 是否千分位展示; 优先于正常精度展示
   bool useZeroPadding = false, // 是否使用0来填充小数部分. 例如0.0000123=>0.0{4}123
   String prefix = '', // 前缀
   String suffix = '', // 后缀
-  String? defIfZero, // 如果为0时的默认展示.
-  String defIfNull = '--', // 如果为空或无效值时的默认展示.
+  String? defIfZero, // 如果为0时的默认展示
+  String defIfNull = '--', // 如果为空或无效值时的默认展示
 }) {
   // 处理数据为空的情况
   if (val == null) return defIfNull;
@@ -38,15 +39,21 @@ String formatNumber(
     return defIfZero;
   }
 
-  String ret = "";
   if (showSign && val >= Decimal.zero) prefix += '+';
 
+  String ret;
   if (showCompact) {
-    ret += val.compact(precision: precision, isClean: cutInvalidZero);
+    final (result, unit) = val.compact(
+      precision: precision,
+      isClean: cutInvalidZero,
+      converter: compactConverter,
+    );
+    ret = result;
+    suffix += unit;
   } else if (showThousands) {
-    ret += val.thousands(precision, mode: mode, isClean: cutInvalidZero);
+    ret = val.thousands(precision, mode: mode, isClean: cutInvalidZero);
   } else {
-    ret += val.formatAsString(precision, mode: mode, isClean: cutInvalidZero);
+    ret = val.formatAsString(precision, mode: mode, isClean: cutInvalidZero);
   }
 
   if (useZeroPadding) {
@@ -58,19 +65,22 @@ String formatNumber(
 
 /// 百分比
 String formatPercentage(
-  double? val, {
+  Decimal? val, {
   int precision = 2,
+  bool showSign = false,
   RoundMode mode = RoundMode.floor,
   bool cutInvalidZero = false,
+  String suffix = '%',
   String defIfNull = '-%', // 如果为空或无效值时的默认展示.
 }) {
   if (val == null) return defIfNull;
   return formatNumber(
-    (val * 100).d,
+    val * hundred,
     precision: precision,
     mode: mode,
+    showSign: showSign,
     cutInvalidZero: cutInvalidZero,
-    suffix: '%',
+    suffix: suffix,
     defIfNull: defIfNull,
   );
 }
@@ -78,7 +88,7 @@ String formatPercentage(
 /// 格式化价钱
 String formatPrice(
   Decimal? val, {
-  required int precision,
+  int precision = 2,
   RoundMode mode = RoundMode.floor,
   bool cutInvalidZero = true,
   bool showThousands = true,
@@ -91,12 +101,12 @@ String formatPrice(
     val,
     precision: precision,
     mode: mode,
-    prefix: prefix,
-    suffix: suffix,
-    defIfZero: defIfZero,
     cutInvalidZero: cutInvalidZero,
     showThousands: showThousands,
     useZeroPadding: useZeroPadding,
+    prefix: prefix,
+    suffix: suffix,
+    defIfZero: defIfZero,
   );
 }
 
@@ -106,15 +116,23 @@ String formatAmount(
   int precision = 2,
   RoundMode? mode,
   bool showCompact = true,
+  CompactConverter? compactConverter,
   bool cutInvalidZero = true,
+  bool useZeroPadding = false,
+  String prefix = '',
+  String suffix = '',
   String? defIfZero,
 }) {
   return formatNumber(
     val,
     precision: precision,
     mode: mode,
-    defIfZero: defIfZero,
-    cutInvalidZero: cutInvalidZero,
     showCompact: showCompact,
+    compactConverter: compactConverter,
+    cutInvalidZero: cutInvalidZero,
+    useZeroPadding: useZeroPadding,
+    prefix: prefix,
+    suffix: suffix,
+    defIfZero: defIfZero,
   );
 }