+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 1.1.0
+- Support Compact Converter.
+- Optimize default parameters.
+
 ## 1.0.0
 - Add number formatting tool.